performance and memory constants too
